修复16进制字符串转数字的问题

+++ src/compiler/Util.java
@@ -15,7 +15,7 @@ public class Util {
         return ints;
     }
 
-    public static long getIntFromStr(String num)
+    public static int getIntFromStr(String num)
     {
         boolean neg = false;
         if(num.startsWith("-"))
@@ -28,12 +28,12 @@ public class Util {
         }
         if(num.startsWith("0x") || num.startsWith("0X"))
         {
-            return Long.parseLong(num.toLowerCase(Locale.ROOT).replace("0x", ""), 16);
+            return (int)Long.parseLong(num.toLowerCase(Locale.ROOT).replace("0x", ""), 16);
         }else if(num.startsWith("0") && num.length()>1)
         {
-            return Long.parseLong(num.toLowerCase(Locale.ROOT).substring(1),8);
+            return (int)Long.parseLong(num.toLowerCase(Locale.ROOT).substring(1),8);
         }
-        long i = Long.parseLong(num);
+        int i = (int)Long.parseLong(num);
         return neg?-i:i;
     }
 }
